requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
scikit-learn>=1.0.0
tqdm>=4.62.0
pillow>=8.0.0
gensim>=4.1.0
matplotlib>=3.4.0
omegaconf>=2.1.0

# Optional dependencies
# sentence-transformers>=2.2.0  # For SBERT embeddings
# fasttext>=0.9.2  # For FastText embeddings
# clearml>=1.6.0  # For experiment tracking 